Fixing test failure on mac by separating cgroup test to linux-specific tests
